requirements.txt
cookies==2.2.1
docopt==0.6.2
funcsigs==0.4
greenlet==0.4.9
mock==1.3.0
pbr==1.8.1
pycorenlp==0.2.0
-e git+https://github.com/exis-io/Exis.git@11025dc9bbcff5186b92dd27aa5ea9275534889b#egg=pyRiffle&subdirectory=python/pyRiffle
python-forecastio==1.3.4
PyYAML==3.11
requests==2.9.1
responses==0.5.1
six==1.10.0
wheel==0.29.0
